+++ src/main/java/com/vonage/client/messages/whatsapp/Whatsapp.java
@@ -17,15 +17,14 @@ package com.vonage.client.messages.whatsapp;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.util.Objects;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 public final class Whatsapp {
 	private final Policy policy;
-	private final String locale;
+	private final Locale locale;
 
-	private Whatsapp(Policy policy, String locale) {
+	private Whatsapp(Policy policy, Locale locale) {
 		this.policy = policy;
 		this.locale = locale;
 	}
@@ -36,15 +35,12 @@ public final class Whatsapp {
 	}
 
 	@JsonProperty("locale")
-	public String getLocale() {
+	public Locale getLocale() {
 		return locale;
 	}
 
-	static Whatsapp construct(Policy policy, String locale) {
+	static Whatsapp construct(Policy policy, Locale locale) {
 		Objects.requireNonNull(locale, "Locale is required");
-		if (locale.length() < 2) {
-			throw new IllegalArgumentException("Invalid locale");
-		}
 		return new Whatsapp(policy, locale);
 	}
 }