+++ src/main/java/com/vonage/client/messages/whatsapp/WhatsappTemplateRequest.java
@@ -46,8 +46,9 @@ public final class WhatsappTemplateRequest extends WhatsappRequest {
 	}
 
 	public static final class Builder extends WhatsappRequest.Builder<WhatsappTemplateRequest, Builder> {
-		String name, locale = "en";
+		String name;
 		List<String> parameters;
+		Locale locale = Locale.ENGLISH;
 		Policy policy;
 
 		Builder() {}
@@ -94,14 +95,12 @@ public final class WhatsappTemplateRequest extends WhatsappRequest {
 
 		/**
 		 * (REQUIRED)
-		 * The BCP 47 language of the template. Defaults to <code>en</code> if not set.
-		 * See <a href=https://developers.facebook.com/docs/whatsapp/api/messages/message-templates>
-		 *     the documentation</a> for supported language options.
+		 * The BCP 47 language of the template. Defaults to {@linkplain Locale#ENGLISH} if not set.
 		 * 
-		 * @param locale The BCP-47 locale.
+		 * @param locale The {@link Locale}.
 		 * @return This builder.
 		 */
-		public Builder locale(String locale) {
+		public Builder locale(Locale locale) {
 			this.locale = locale;
 			return this;
 		}